Module packed_struct::derive

The derivation macros for packing and enums.

Derive Macros

PackedStruct
The derive macro that generates the packing and unpacking code for your structure.

PrimitiveEnum
A derive macro that generates packing and unpacking code for simple enum variants. It helps with converting your enums into integer types and back, with many other helper traits.

PrimitiveEnum_i8

PrimitiveEnum_i16

PrimitiveEnum_i32

PrimitiveEnum_i64

PrimitiveEnum_u8

PrimitiveEnum_u16

PrimitiveEnum_u32

PrimitiveEnum_u64