Module packed_struct::prelude

Re-exports the most useful traits and types. Meant to be glob imported.

Re-exports

pub use super::derive::*;

pub use crate::types::*;

pub use crate::types::bits as packed_bits;

Enums

EnumCatchAll
A wrapper for primitive enums that supports catching and retaining any values that don’t have defined discriminants.

PackingError
Packing errors that might occur during packing or unpacking

Traits

PackedStruct
A structure that can be packed and unpacked from a byte array.

PackedStructSlice
A structure that can be packed and unpacked from a slice of bytes.

PrimitiveEnum
An enum type that can be packed or unpacked from a simple primitive integer.

PrimitiveEnumStaticStr
Static display formatters.