Trait packed_struct::prelude::PackedStructSlice

pub trait PackedStructSlice where
    Self: Sized, {
    fn pack_to_slice(&self, output: &mut [u8]) -> PackingResult<()>;

    fn unpack_from_slice(src: &[u8]) -> PackingResult<Self>;

    fn packed_bytes_size(opt_self: Option<&Self>) -> PackingResult<usize>;
}

A structure that can be packed and unpacked from a slice of bytes.

Required methods

fn pack_to_slice(&self, output: &mut [u8]) -> PackingResult<()>

Pack the structure into an output buffer.

fn unpack_from_slice(src: &[u8]) -> PackingResult<Self>

Unpack the structure from a buffer.

fn packed_bytes_size(opt_self: Option<&Self>) -> PackingResult<usize>

Number of bytes that the type or this particular instance of this structure demands for packing or unpacking.

Loading content...