Trait packed_struct::prelude::PrimitiveEnum

pub trait PrimitiveEnum where
    Self: Sized + Copy, {
    type Primitive: PartialEq + Sized + Copy + Debug;
    fn from_primitive(val: Self::Primitive) -> Option<Self>;

    fn to_primitive(&self) -> Self::Primitive;

    fn from_str(s: &str) -> Option<Self>;

    fn from_str_lower(s: &str) -> Option<Self>;
}

An enum type that can be packed or unpacked from a simple primitive integer.

Associated Types

type Primitive: PartialEq + Sized + Copy + Debug

The primitve type into which we serialize and deserialize ourselves.

Loading content...
Required methods

fn from_primitive(val: Self::Primitive) -> Option<Self>

Convert from a primitive, might fail.

fn to_primitive(&self) -> Self::Primitive

Convert to a primitive value.

fn from_str(s: &str) -> Option<Self>

Convert from a string value representing the variant. Case sensitive.

fn from_str_lower(s: &str) -> Option<Self>

Convert from a string value representing the variant. Lowercase.

Loading content...