Trait packed_struct::PackedStructInfo

pub trait PackedStructInfo {
    fn packed_bits() -> usize;
}

Infos about a particular type that can be packaged.

Required methods

fn packed_bits() -> usize

Number of bits that this structure occupies when being packed.

Loading content...