Module packed_struct::types

Implementations and wrappers for various packing types.

Modules

bits
Types that specify the exact number of bits a packed integer should occupy.

Structs

BitOne

BitZero

Integer
A bit-limited integer, stored in a native type that is at least as many bits wide as the desired size.

LsbInteger
A wrapper that packages the integer as a LSB packaged byte array. Usually invoked using code generation.

MsbInteger
A wrapper that packages the integer as a MSB packaged byte array. Usually invoked using code generation.

ReservedBits
Always packs into the associated bit value. Ignores the input when unpacking.

Traits

IntegerAsBytes
Convert a native platform integer type into a byte array.

ReservedBitValue

SizedInteger
Convert an integer of a specific bit width into native types.

SizedIntegerSigned
A helper for converting specific bit width signed integers into the native type.

Type Definitions

ReservedOne
Packs into a set of ones. Ignores the input when unpacking.

ReservedOnes

ReservedZero
Packs into a set of zeroes. Ignores the input when unpacking.

ReservedZeroes