Module packed_struct::types::bits

Types that specify the exact number of bits a packed integer should occupy.

Structs

Bits1

Bits2

Bits3

Bits4

Bits5

Bits6

Bits7

Bits8

Bits9

Bits10

Bits11

Bits12

Bits13

Bits14

Bits15

Bits16

Bits17

Bits18

Bits19

Bits20

Bits21

Bits22

Bits23

Bits24

Bits25

Bits26

Bits27

Bits28

Bits29

Bits30

Bits31

Bits32

Bits33

Bits34

Bits35

Bits36

Bits37

Bits38

Bits39

Bits40

Bits41

Bits42

Bits43

Bits44

Bits45

Bits46

Bits47

Bits48

Bits49

Bits50

Bits51

Bits52

Bits53

Bits54

Bits55

Bits56

Bits57

Bits58

Bits59

Bits60

Bits61

Bits62

Bits63

Bits64

Bits65

Bits66

Bits67

Bits68

Bits69

Bits70

Bits71

Bits72

Bits73

Bits74

Bits75

Bits76

Bits77

Bits78

Bits79

Bits80

Bits81

Bits82

Bits83

Bits84

Bits85

Bits86

Bits87

Bits88

Bits89

Bits90

Bits91

Bits92

Bits93

Bits94

Bits95

Bits96

Bits97

Bits98

Bits99

Bits100

Bits101

Bits102

Bits103

Bits104

Bits105

Bits106

Bits107

Bits108

Bits109

Bits110

Bits111

Bits112

Bits113

Bits114

Bits115

Bits116

Bits117

Bits118

Bits119

Bits120

Bits121

Bits122

Bits123

Bits124

Bits125

Bits126

Bits127

Bits128

Bits129

Bits130

Bits131

Bits132

Bits133

Bits134

Bits135

Bits136

Bits137

Bits138

Bits139

Bits140

Bits141

Bits142

Bits143

Bits144

Bits145

Bits146

Bits147

Bits148

Bits149

Bits150

Bits151

Bits152

Bits153

Bits154

Bits155

Bits156

Bits157

Bits158

Bits159

Bits160

Bits161

Bits162

Bits163

Bits164

Bits165

Bits166

Bits167

Bits168

Bits169

Bits170

Bits171

Bits172

Bits173

Bits174

Bits175

Bits176

Bits177

Bits178

Bits179

Bits180

Bits181

Bits182

Bits183

Bits184

Bits185

Bits186

Bits187

Bits188

Bits189

Bits190

Bits191

Bits192

Bits193

Bits194

Bits195

Bits196

Bits197

Bits198

Bits199

Bits200

Bits201

Bits202

Bits203

Bits204

Bits205

Bits206

Bits207

Bits208

Bits209

Bits210

Bits211

Bits212

Bits213

Bits214

Bits215

Bits216

Bits217

Bits218

Bits219

Bits220

Bits221

Bits222

Bits223

Bits224

Bits225

Bits226

Bits227

Bits228

Bits229

Bits230

Bits231

Bits232

Bits233

Bits234

Bits235

Bits236

Bits237

Bits238

Bits239

Bits240

Bits241

Bits242

Bits243

Bits244

Bits245

Bits246

Bits247

Bits248

Bits249

Bits250

Bits251

Bits252

Bits253

Bits254

Bits255

Bytes0

Bytes1

Bytes2

Bytes3

Bytes4

Bytes5

Bytes6

Bytes7

Bytes8

Bytes9

Bytes10

Bytes11

Bytes12

Bytes13

Bytes14

Bytes15

Bytes16

Bytes17

Bytes18

Bytes19

Bytes20

Bytes21

Bytes22

Bytes23

Bytes24

Bytes25

Bytes26

Bytes27

Bytes28

Bytes29

Bytes30

Bytes31

Bytes32

Traits

BitsFullBytes
These bits are a multiple of 8

BitsPartialBytes
These bits are not a multiple of 8

ByteArray
Helper that allows us to cast a fixed size array into a byte slice.

NumberOfBits
Number of bits that the generic type should occupy.

NumberOfBytes
Number of bytes that the generic type should occupy.