Trait packed_struct::types::bits::NumberOfBits

pub trait NumberOfBits: Copy + Clone + Debug + Default {
    type Bytes: NumberOfBytes;
    fn number_of_bits() -> usize;
}

Number of bits that the generic type should occupy.

Associated Types

type Bytes: NumberOfBytes

Minimal number of bytes that this bit width requires.

Loading content...
Required methods

fn number_of_bits() -> usize

The numerical number of bits.

Loading content...