Trait packed_struct::types::bits::NumberOfBytes

pub trait NumberOfBytes: Copy + Clone + Debug + Default {
    type AsBytes: ByteArray;
    fn number_of_bytes() -> usize;
}

Number of bytes that the generic type should occupy.

Associated Types

type AsBytes: ByteArray

The byte array type that holds these bytes, for instance [u8; 2].

Loading content...
Required methods

fn number_of_bytes() -> usize

The numberical number of bytes.

Loading content...