Struct packed_struct::types::Integer

pub struct Integer<T, B> { /* fields omitted */ }

A bit-limited integer, stored in a native type that is at least as many bits wide as the desired size.

Implementations

impl<T, B> Integer<T, B> where
Self: Copy,

pub fn as_packed_msb(&self) -> MsbInteger<T, B, Self>

Convert into a MSB packing helper

pub fn as_packed_lsb(&self) -> LsbInteger<T, B, Self>

Convert into a LSB packing helper