Trait packed_struct::types::SizedIntegerSigned

pub trait SizedIntegerSigned<T, B>: SizedInteger<T, B> where
    B: NumberOfBits, {
    fn from_unpacked_to_signed(val: T) -> T;
}

A helper for converting specific bit width signed integers into the native type.

Required methods

fn from_unpacked_to_signed(val: T) -> T

Sign-extends the packed value into a properly signed representation in one’s complement.

Loading content...